test: consolidate the four independent makeBSplineSurface() fixtures (#1254)
BSplineSurfaceManipulationTests, BSplineSurfaceExtrasTests,
BSplineSurfaceRemoveVKnotTests and BSplineSurfaceCompletionsV121Tests each
independently reimplemented a private `makeBSplineSurface()` helper, four
genuinely different fixtures reusing the same name in different files.

Moved each construction into SurfaceTestFixtures.swift under a distinct,
descriptive name (makeCylinderDerivedBSplineSurface,
makeModThreeGridBSplineSurface, makeSinCosGridBSplineSurface,
makeExplicitPoleBSplineSurface), matching the shared-fixture home this
directory already established for the #645 Gordon-network fixture. Each
suite's own makeBSplineSurface() is now a one-line forward to the shared
function, so every existing call site is unchanged. Zero behavior change.
